requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
google-generativeai
Pillow
python-decouple
fastapi
uvicorn[standard]
python-dotenv
pytesseract
Pillow
openai
google-generativeai
requests